feat: improve package bundling
