+++ .editorconfig
@@ -1,18 +1,18 @@
+# https://editorconfig.org
 root = true
 
 [*]
-end_of_line = lf
-insert_final_newline = true
-trim_trailing_whitespace = true
 charset = utf-8
+end_of_line = lf
+indent_size = 2
+indent_style = space
+insert_final_newline = true
+max_line_length = 80
+trim_trailing_whitespace = true
 
 [*.md]
+max_line_length = 0
 trim_trailing_whitespace = false
 
-[*.{js,json}]
-indent_style = tab
-indent_size = 2
-
-[*.yml]
-indent_style = space
-indent_size = 2
+[COMMIT_EDITMSG]
+max_line_length = 0
