+++ src/server/utils/prototype-injection-handler.utils.js
@@ -1,26 +1,28 @@
 const Bourne = require('@hapi/bourne');
 const { resolveSchema } = require('./schema.utils');
-// eslint-disable-next-line no-undef
-module.exports = prototypeInjectionHandler = (req, res, next) => {
-	try {
-		if (req.body && typeof req.body === 'object') {
-			Bourne.scan(req.body);
-		}
-		next();
-	} catch (error) {
-		let OperationOutcome = resolveSchema('3_0_1', 'operationoutcome');
-		let err = new OperationOutcome({
-			statusCode: 400,
-			issue: [
-				{
-					severity: 'error',
-					code: 'invalid',
-					details: {
-						text: error.message,
-					},
-				},
-			],
-		});
-		next(err);
-	}
+
+const prototypeInjectionHandler = (req, res, next) => {
+  try {
+    if (req.body && typeof req.body === 'object') {
+      Bourne.scan(req.body);
+    }
+    next();
+  } catch (error) {
+    let OperationOutcome = resolveSchema('3_0_1', 'operationoutcome');
+    let err = new OperationOutcome({
+      statusCode: 400,
+      issue: [
+        {
+          severity: 'error',
+          code: 'invalid',
+          details: {
+            text: error.message,
+          },
+        },
+      ],
+    });
+    next(err);
+  }
 };
+
+module.exports = prototypeInjectionHandler;
